+++ src/compiler/pt/method_call.h
@@ -41,7 +41,7 @@ namespace g0at
             std::wstring get_right() { return right; }
             std::wstring get_name() { return right; }
             void add_arg(lib::pointer<expression> arg) { args.push_back(arg); }
-            int get_args_count() { return args.size(); }
+            int get_args_count() { return (int)args.size(); }
             lib::pointer<expression> get_arg(int index) { return args.at(index); }
 
         protected:
